redesign_app.py
import base64
import io
import os
import time
import json

import numpy as np
import dash
import dash_core_components as dcc
import dash_html_components as html
from PIL import Image
from io import BytesIO
from dash.dependencies import Input, Output, State
import pandas as pd
import plotly.graph_objs as go
import scipy.spatial.distance as spatial_distance
